finished off backtest

+++ PerformancePlaci/Data/Line.cs
@@ -6,6 +6,6 @@ namespace PerformancePlaci.data
     public class Line
     {
         public DateTime Date { get; set; }
-        public List<float> Values { get; set; } = new List<float>();
+        public List<double> Values { get; set; } = new List<double>();
     }
 }
